code merged

+++ app/(main_layout)/(dashboard_layout)/users/Content.tsx
@@ -9,9 +9,10 @@ import Modal from "@/components/Modal";
 import { deleteApi, getApi, postApi, putApi } from "@/lib/apiClient";
 import DebounceInput from "@/components/DebounceInput";
 import useZStore from "@/hooks/store";
+import swal from "@/components/Swal";
 
 export default function Content() {
-  const {user,setUser} = useZStore()
+  const { user, setUser } = useZStore()
   const [editId, setEditId] = useState<string | null>(null);
   const [formModal, setFormModal] = useState<any>();
   const [data, setData] = useState<any[]>([]);
@@ -31,7 +32,7 @@ export default function Content() {
     if (editId) {
       res = await putApi({ url: '/api/users', payload: { id: editId, name: formModal.name, email: formModal.email } })
     } else {
-      res = await postApi({ url: '/api/users', payload: { name: formModal.name, email: formModal.email ,password:formModal.password} })
+      res = await postApi({ url: '/api/users', payload: { name: formModal.name, email: formModal.email, password: formModal.password } })
     }
     setFormLoading(false)
     if (res.success) {
@@ -50,9 +51,13 @@ export default function Content() {
   };
 
   const deleteUser = (id: string) => {
-    deleteApi({ url: `api/users/${id}` }).then(res => {
-      if (res.success) {
-        clearFilter()
+    swal({ title: `Do you want to delete "${data?.find(itm=>itm._id)?.name}"`, showCancel: true,icon:'warning' }).then(res => {
+      if (res.isConfirmed) {
+        deleteApi({ url: `api/users/${id}` }).then(res => {
+          if (res.success) {
+            clearFilter()
+          }
+        })
       }
     })
   }
@@ -78,9 +83,7 @@ export default function Content() {
       key: 'actions',
       render: (item: any) => <>
         <button onClick={() => handleEdit(item)} className="mr-2 text-blue-500 cursor-pointer">Edit</button>
-        <button onClick={async () => {
-          await deleteUser(item._id)
-        }} className="text-red-500 cursor-pointer">Delete</button>
+        <button onClick={() => deleteUser(item._id)} className="text-red-500 cursor-pointer">Delete</button>
       </>
     }
   ]
@@ -133,7 +136,7 @@ export default function Content() {
       if (apiRef.current) apiRef.current?.abort()
     }
   }, [])
-  
+
   return (
     <div>
       <div className="container mx-auto px-4 py-8">
@@ -233,15 +236,15 @@ export default function Content() {
               onChange={e => setFormModal((prev: any) => ({ ...prev, email: e.target.value }))}
               required className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:border-transparent" />
           </div>
-          {!editId?<>
-          <div>
-            <label htmlFor="userName" className="block text-sm font-medium text-gray-700 mb-1">Password</label>
-            <input type="text"
-              value={formModal.password}
-              onChange={e => setFormModal((prev: any) => ({ ...prev, password: e.target.value }))}
-              required minLength={8} className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:border-transparent" />
-          </div>
-          </>:<></>}
+          {!editId ? <>
+            <div>
+              <label htmlFor="userName" className="block text-sm font-medium text-gray-700 mb-1">Password</label>
+              <input type="text"
+                value={formModal.password}
+                onChange={e => setFormModal((prev: any) => ({ ...prev, password: e.target.value }))}
+                required minLength={8} className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:border-transparent" />
+            </div>
+          </> : <></>}
           <div className="flex justify-end space-x-3">
             <button type="button" className="px-5 py-2.5 border border-gray-300 rounded-lg text-gray-700 hover:bg-gray-50"
               onClick={() => setFormModal('')}
